予約変更ダイアログ、録画結果情報ダイアログに「前へ」「次へ」ボタンを追加
・予約変更ダイアログには「操作後も閉じない」オプションを追加し、
　オプション有効時のみ「前へ」「次へ」と関連UIを表示するようにした。
・録画結果情報ダイアログは、ダイアログ側からの削除などを可能とした。
・使用予定チューナー画面及び番組表で、Ctrl+上下キーで、　予約データをたどれるようにした。
　(ダイアログの「前へ」「次へ」ボタンのショートはCtrl+左右に変更)

+++ EpgTimer/EpgTimer/UserCtrlView/ListBoxDragMoverView.xaml.cs
@@ -43,8 +43,8 @@ namespace EpgTimer.UserCtrlView
 
         public abstract class LVDMHelper
         {
-            public abstract uint GetID(object data);
-            public virtual bool SaveChange(Dictionary<uint, uint> changeIDTable) { return true; }
+            public abstract ulong GetID(object data);
+            public virtual bool SaveChange(Dictionary<ulong, ulong> changeIDTable) { return true; }
             public virtual bool RestoreOrder() { return true; }
             public virtual void StatusChanged() { }
             public virtual void ItemMoved() { }//アイテム動かなくてもステータス変更の場合がある
@@ -177,13 +177,13 @@ namespace EpgTimer.UserCtrlView
             }
             catch (Exception ex) { MessageBox.Show(ex.Message + "\r\n" + ex.StackTrace); }
         }
-        public Dictionary<uint, uint> GetChangeIDTable()
+        public Dictionary<ulong, ulong> GetChangeIDTable()
         {
-            List<uint> dataIdList = this.dataList.OfType<object>().Select(item => hlp.GetID(item)).ToList();
+            List<ulong> dataIdList = this.dataList.OfType<object>().Select(item => hlp.GetID(item)).ToList();
             dataIdList.Sort();//存在する(再利用可能な)IDリストを若い順(元の順)に並べたもの。
 
             //並び替えテーブル
-            var changeIDTable = new Dictionary<uint, uint>();
+            var changeIDTable = new Dictionary<ulong, ulong>();
             for (int i = 0; i < this.dataList.Count; i++)
             {
                 changeIDTable.Add(hlp.GetID(this.dataList[i]), dataIdList[i]);
